refactor: split on-chain Scalus scripts into cardano-onchain sbt subproject
Moves all @Compile objects (validators, state types, extension objects)
into a new `cardano-onchain/` sbt subproject so on-chain code is isolated
from the core application.

Key changes:
- New `cardanoOnchain` sbt project with its own `scalus-plugin_3` compiler
  plugin entry (required; plugin is not inherited by subprojects)
- `EvacuationKey` moves to cardano-onchain; circe/collection givens stay in
  core as `EvacuationMapInstances`
- `TreasuryState` on-chain types move to cardano-onchain; off-chain
  conversion methods extracted to `TreasuryStateConversions` object in core
- `StandaloneEvacuationCommitmentOnchain` `FromData`/`ToData` givens moved
  inside `@Compile object VoteState` (top-level `@Compile case class`
  triggers Scalus EmptyTree on constructor fields)
- `ByteStringExtension` loses `@Compile` (inline-only objects cannot have
  SIR annotations; Scala 3 inlines them at call sites instead)
