+++ cardano-onchain/src/main/scala/hydrozoa/lib/cardano/scalus/ledger/api/ByteStringExtensions.scala
@@ -1,10 +1,8 @@
 package hydrozoa.lib.cardano.scalus.cardano.onchain.plutus
 
-import scalus.*
 import scalus.uplc.builtin.Builtins.*
 import scalus.uplc.builtin.{Builtins, ByteString}
 
-@Compile
 object ByteStringExtension:
     extension (self: ByteString)
         inline infix def <(other: ByteString): Boolean =
