+++ cardano-onchain/src/main/scala/hydrozoa/rulebased/ledger/l1/script/plutus/RuleBasedTreasuryScript.scala
